Use MockReplicaStateMachine and refactor tests
